+++ src/main/java/racingcar/domain/numbergenerator/NumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.numbergenerator;
 
 public interface NumberGenerator {
     int generate();
