+++ src/main/java/racingcar/domain/racing/NumberApplicatorToCar.java
@@ -1,6 +1,9 @@
-package racingcar.domain;
+package racingcar.domain.racing;
 
 import java.util.List;
+import racingcar.domain.Car;
+import racingcar.domain.CarRepository;
+import racingcar.domain.numbergenerator.NumberGenerator;
 
 
 public class NumberApplicatorToCar {
