+++ src/test/java/racingcar/domain/numbergenerator/FixedNumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.numbergenerator;
 
 
 public class FixedNumberGenerator implements NumberGenerator {
